Перенос функция отбражения попапов с ошибкой и успешной отправкой в отдельный модуль
