+++ js/form.js
@@ -2,6 +2,8 @@ import { isEscapeKey, showAlert} from './util.js';
 import {addScaleActive, removeScaleActive} from './scale.js';
 import {addEffectActive, removeEffectActive} from './effects.js';
 import {sendData} from './api.js';
+import { addError, errorEscKeydown, addSuccess } from './error-success.js';
+
 
 const MAX_HASHTAG_COUNT = 5;
 const HASHTAG_SYMBOLS = /^#[a-zа-яё0-9]{1,19}$/i;
@@ -14,8 +16,6 @@ const loadFileField = form.querySelector('#upload-file');
 const hashTagField = overlay.querySelector('.text__hashtags');
 const descriptionField = overlay.querySelector('.text__description');
 const crossButton = overlay.querySelector('.img-upload__cancel');
-const errorTemplate = document.querySelector('#error').content.querySelector('.error');
-const successTemplate = document.querySelector('#success').content.querySelector('.success');
 const submitButton = form.querySelector('.img-upload__submit');
 
 const pristine = new Pristine(form, {
@@ -117,49 +117,8 @@ pristine.addValidator(
 
 loadFileField.addEventListener('change', onLoadFileFieldClick);
 
-const errorElement = errorTemplate.cloneNode(true);
-const successElement = successTemplate.cloneNode(true);
 
-const onErrorEscKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    document.addEventListener('keydown', onFormEscKeydown);
-    errorElement.remove();
-    document.removeEventListener('keydown', onErrorEscKeydown);
-  }
-};
-
-const addError = () => {
-  document.removeEventListener('keydown', onFormEscKeydown);
-  document.addEventListener('keydown', onErrorEscKeydown);
-  const errorButton = errorElement.querySelector('.error__button');
-  errorButton.addEventListener('click', () => {
-    errorElement.remove();
-    document.removeEventListener('keydown', onErrorEscKeydown);
-    document.addEventListener('keydown', onFormEscKeydown);
-  });
-  body.appendChild(errorElement);
-};
-
-const onSuccessEscKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    successElement.remove();
-    document.removeEventListener('keydown', onSuccessEscKeydown);
-  }
-};
-
-const addSuccess = () => {
-  document.addEventListener('keydown', onSuccessEscKeydown);
-  const successButton = successElement.querySelector('.success__button');
-  successButton.addEventListener('click', () => {
-    successElement.remove();
-    document.removeEventListener('keydown', onSuccessEscKeydown);
-  });
-  body.appendChild(successElement);
-};
-
-
+// Отправка формы на сервер
 const SubmitButtonText = {
   IDLE: 'Опубликовать',
   SENDING: 'Пубилкую...'
@@ -175,6 +134,8 @@ const unblockSubmitButton = () => {
   submitButton.textContent = SubmitButtonText.IDLE;
 };
 
+const onErrorEscKeydown = errorEscKeydown(onFormEscKeydown);
+
 const setUserFormSubmit = (onSuccess) => {
   form.addEventListener('submit', (evt) => {
     evt.preventDefault();
@@ -184,13 +145,13 @@ const setUserFormSubmit = (onSuccess) => {
       blockSubmitButton();
       sendData(new FormData(evt.target))
         .then(onSuccess)
-        .then(addSuccess())
+        .then(addSuccess)
         .catch((err) => {
           showAlert(err.message);
         })
         .finally(unblockSubmitButton);
     } else {
-      addError();
+      addError(onFormEscKeydown, onErrorEscKeydown);
     }
   });
 };
